requirements.txt
streamlit
torch
transformers[torch]
Pillow